backend/pyproject.toml
[project]
name = "medtrace-backend"
version = "0.1.0"
description = "Backend FastAPI for MedTrace"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "sqlalchemy>=2.0.38",
  "psycopg[binary]>=3.2.6",
  "alembic>=1.16.4",
  "pydantic>=2.10.6",
  "pydantic-settings>=2.8.1",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "python-multipart>=0.0.20",
  "email-validator>=2.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
  "httpx>=0.28.1",
]

[build-system]
requires = ["setuptools>=75.8.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
